Update 20180824-uk-pilot-topicmodels.R
Hello, I add here a few comments with the structure that Michael proposed in his summary of our meeting last Friday. 

Control variables: there are already more than enough controls. Having the entity name in the group might be redundant if we have already a dummy that identified problematic cases. On the correlation between variables and types of governors, the kind of problems pointed out by Michael are going to be present throughout. Each IP will dominate in one particular territorial_scope and policy_scope. For instance, all of IP2 is going to be actor_type=public and none is going to be "regional_national". However, I do not think we need to exclude anything from the model, because as long as we include entity name as a control variable and we can analyze considering the effect of IPs, keeping all variables in the model is how we are able to make general statements.

Dummy variables to identify particular governors: In my case, it is important to point out the EC, with #entity_id 304219, and entity_name european commission. 

Recoding of the independent variables: I checked the recoding according to my IP, see my branch. Only a couple of comments, I think we are looking at the right categories already.